requirements.txt
openai
requests
urllib3
python-dotenv